fix: resolve SIGSEGV on Linux for sqlite_bind_int64
The issue was caused by ABI incompatibility when mixing 32-bit (Int) and
64-bit (Int64) parameters in C FFI calls. On Linux x86-64, MoonBit's
parameter passing for mixed-width integers differed from what the C
compiler expected.

Solution:
- Changed C stub to accept idx as int64_t (cast back to int internally)
- Added internal FFI function sqlite_bind_int64_ffi with Int64 parameters
- Wrapper function sqlite_bind_int64 maintains backward-compatible API

Fixes #6

+++ stub.c
@@ -153,8 +153,11 @@ int32_t sqlite_bind_null(sqlite3_stmt* stmt, int32_t idx) {
     return sqlite3_bind_null(stmt, idx);
 }
 
-int32_t sqlite_bind_int64(sqlite3_stmt* stmt, int32_t idx, int64_t value) {
-    return sqlite3_bind_int64(stmt, idx, value);
+// Note: Using int64_t for idx to avoid ABI issues with mixed 32/64-bit parameters
+// on some platforms (Linux x86-64). MoonBit may pass parameters differently
+// when mixing Int and Int64 types.
+int32_t sqlite_bind_int64(sqlite3_stmt* stmt, int64_t idx, int64_t value) {
+    return sqlite3_bind_int64(stmt, (int)idx, value);
 }
 
 int32_t sqlite_bind_blob(sqlite3_stmt* stmt, int32_t idx, moonbit_bytes_t blob) {
